Author field. Use '---' separator for simplecms
The load() code could be refactored to look like simplebook, but at
least the format should be the same.

Author field is a nice thing to have for future growth to multiple
author websites (I know that the VCS logs will have all of this, but
this will have it as part of the webpage.

"Webpages by ..." could be a nice feature.